Added /t unclaim outpost command
- The /t unclaim outpost command does following:
It checks the arguments when in the AreaSelectionUtil.
If the value is outpost in the arguments it checks against the town
if the townblock really is an outpost.
If it is, it sets the outpost value to true and returns the townblock the
player is on (standard behavior for unclaiming a townblock)
since the code for deleting a townblock does remove the outpost itself from
the town.