Add custom Metadata to Towns and Townblocks (#3482)
## This pull request aims to add custom metadata to towny.

### Metadata implementation
Each metadata object is a subclass of the abstract generic class `CustomFieldData<T>`
These subclasses include:
- `IntegerDataField` - `CustomFieldData<Integer>`
- `StringDataField`- `CustomFieldData<String>`
- `BooleanDataField` - `CustomFieldData<Boolean>`
- `DecimalDataField`- `CustomFieldData<Double>`

Metadata is hooked up to both `Town` objects and `TownBlock` objects, and they are lazy loaded so they don't effect memory usage if never used.
Metadata is stored in the text file or sql with the new `metadata` attribute. This attribute contains the key, value, and type in order to perform proper loading of the object.

### Metadata will be registered by a plugin developer preferably in the onLoad() function.
```
IntegerDataField minLevel = new IntegerDataField("min-level", 30); // Sets the default value for this field to 30 
TownyAPI.getInstance().registerCustomMetaData(minLevel);
```

These objects will be stored in the TownyUniverse singleton, in order to check for duplication.

Metadata can then be accessed through the `Town` or `Townblock`'s new getMetadata() getter, which returns a hash set of all of the meta data associated with the object. From there it is up to the developer to save the object when any changes are made.

### In-Game Metadata manipulation
When working with other admins it's nice to be able to rapidly make changes dynamically without relying on the parent plugin. So in-game commands to modify metadata for Towns and TownBlocks are available:

- `/ta (town {town} or plot) meta add {keyname}`
- `/ta (town {town} or plot) meta set {keyname} {value}`
- `/ta (town {town} or plot) meta remove {keyname}`

#### Note that keyname must be the name of a key registered by a plugin