Automation: Javadocs @ Deploy Phase & GitHub Actions
Javadocs now only build at the Maven 'deploy' phase, when the jarfile is built and pushed to the GitHub Package Registry by a CI.   

Normal users can still build JavaDocs manually using `mvn javadocs:jar` from a terminal, or by selecting the plugin in IntelliJ's sidebar.  

This commit also introduces a GitHub Actions workflow file that automatically builds upon a Release or Prerelease - which will push to the Package Registry for us. (As I don't have contributor-level access, this may or may not work out of the box... Trial and error.)