Added support for multiple currencies.
To configure, set plugin.interfacing.currency_name to something other then
a blank string.

Formatting is a bit spotty still, due to a bug in TNE, when an explicit
currency is provided through Reserve, formatting does not work. When
not using an explicit currency, or when using Vault, formatting works
fine.