Fix mayoral succession rank ordering issue
Fix issue where mayoral succession rank loop not being broken/exited
when a new mayor is appointed, which had resulted in the first eligible
resident for each rank in the order of succession being being made mayor
(with the last one to be made mayor ending up as the mayor when the
succession process completed).

Update config comment for the order of mayoral succession to reflect the
rank ordering issue being resolved.