Refactor/nations dont save towns anymore (#4261)
* Removes towns.txt and nations.txt files.

Removes necessity of saving the towns and nations lists.
Loads old towns.txt and nations.txt to ensure that orphaned files in the
towns & nations folders are not loaded. If successfully loaded, the
towns.txt and nations.txt files are deleted.

* Refactor for Towns loading their mayors.

If a town attempts to set a mayor who is not part of the town, and the
town has no residents, the town will be deleted.
  - Closes #4170.

* Refactor removing nations storing their towns.

Workload of adding/removing a nation for a town now worked out in Town
instead of Nation.
Reduces all but one EmptyNationException.

Untested.

* Slight tweaking on initial commit.

* Clean up nation/town loading, handle loading nation capitals
intelligently.

* Re-work mayor setting. Find new mayors when a town loads a mayor who
doesn't match.

TODO: testing.

* Improve readability in Nation.java's findNewCapital().

* Add missing return to forceSetCapital(Town capital).

* Merge master branch and sort out conflicts in TownySQLSource.

Also fix town loading after selecting a new mayor.