Ensure Objects are properly assigned, closed, and compared to improve safety and accuracy. (#4496)
* Fix issues with object comparisons.

- Convert  `node == "string"` to `equals(string)`
  - Was not comparing actual value but locations in memory.

- Fix incompatible objects always equaling false.
  - Was comparing incompatible types.

- Use negated equals(), not `!=` operator
  - Again, comparing contents, not numbers or locations in memory.

* Cast something to a double/long in order to actually get a double/long

When arithmetic is performed on integers, the result will always be an integer. You can assign that result to a long, double, or float with automatic type conversion...

<https://wiki.sei.cmu.edu/confluence/display/java/NUM50-J.+Convert+integers+to+floating+point+for+floating-point+operations>

* Increase precision of BigDecimal

Because of floating point imprecision, you're unlikely to get the value you expect from the BigDecimal(double) constructor.
<https://wiki.sei.cmu.edu/confluence/display/java/NUM10-J.+Do+not+construct+BigDecimal+objects+from+floating-point+literals>

* Don't ignore return value of replaceAll()

What is the point of making the call if the results are ignored?

See <https://wiki.sei.cmu.edu/confluence/display/java/EXP00-J.+Do+not+ignore+values+returned+by+methods> for details.

* Resources should be closed.

Connections, streams, files, and other classes that implement the Closeable interface or its super-interface, AutoCloseable, needs to be closed after use. Further, that close call must be made in a finally block otherwise an exception could keep the call from being made. Preferably, when class implements AutoCloseable, resource should be created using "try-with-resources" pattern and will be closed automatically.

Failure to properly close resources will result in a resource leak which could bring first the application and then perhaps the box the application is on to their knees.

* Unused import and unneeded warning removed.

Co-authored-by: Llm Dl <LlmDlio@gmail.com>