Re-pkg WarPreUnclaimEvent as TownPreUnclaimCmdEvent in `event.town` pkg.
Makes a bit more sense, when you think about it. It's capable of being called by more than just war plugins, and fragmenting events into separate packages within the same project is insane on it's own.

Added note to class Javadoc about potential confusion with TownyPreUnclaimEvent.

Thanks to @Silverwolfg11 and @LlmDl for discussing this and offering the suggestion.