Remove EconomyException from being thrown by getHoldingBalance.
After some investigating into the history of the NoClassDefFoundError
being caught in the getHoldingBalance() method in Account, which then
results in an EconomyException, I found that it has been in use going
back at least 10 years.

getHoldingBalance used to reside in the EconomyAccount class, where upon
every use it would try to match either Register or iConomy directly (no
Vault or Reserve back then,) and if it failed to find either plugin it
would throw the NoClassDefFoundError.

The same try/catch has been dragged forwards through the years even
though everything else has changed around it. These days before we do
anything economy-related we are checking if the
TownyEconomyHandler.isActive() and if it is there should be no
possiblity to ever catch a NoClassDefFoundError.