Remove Integrated FlagWar (admins: use add-on)
The integrated version of FlagWar has been deprecated, and marked for removal, since 0.97.0.0.

If you require FlagWar, we have moved it to it's own add-on. You can find it at https://github.com/TownyAdvanced/FlagWar

Note: The FlagWar add-on has some new features, but may also introduce some bugs not found in the integrated version. Please report any bugs for FlagWar to it's own issue tracker.