Remove the GatherResidentUUIDTask
The API method we used to gather a UUID from a username was apparently
removed (but the removal was not documented.) This makes the entire task
useless, which is why it is being removed. This PR removes this task,
which acted as a way for databases to be filled in.

Ultimately, between players having their UUIDs recorded when they log
in, and the old-resident-delete feature, 96.7% of servers are reporting
UUIDs stored for 90% or more residents.

https://bstats.org/plugin/bukkit/Towny/2244

This PR leaves behind a couple of methods in BukkitTools which are
unused but could be useful.