Add ally and enemy lists to Towns.
These are persistant entries in the database, which populate enemy and
ally lists for towns. The lists are viewable via `/t allylist
{townname}` and `/t enemylist {townname}`. The CombatUtil has been
modified to account for towns being enemied.

So far, these are used in EventWar for the TownWar and CivilWar war
types.

Drawbacks/design choices:

There is no way to set enemy and ally lists in game however, it is all
handled via the API. I am fairly happy with this limitation, I feel
exposing it ingame would take away from nations too much.

There is currently no events being thrown to prevent one plugin from
modifying the choices of another plugin adding or removing a ally or
enemy from a town. This is something I could see happening before this
PR is merged.