Queue the revert on unclaim feature.
Stops Towny from loading the entire regen queue into memory and
processing all the reverts at the same time.

Towny not loads the regen.txt into a list of WorldCoords which holds all
active and inactive revert-on-unclaims.

When there are less than 20 active reverts, Towny will look to see if
there are any available WorldCoords in the queue. If one is found only
then will the PlotBlockData object be created and loaded into memory.