Fix AreaSelectionUtil treating every selection like it will be for claiming.
Adds claim boolean to the selectWorldCoordArea, so that the available
selection size is only limited when the selection is actually for
claiming.

Selecting areas for unclaiming, and selecting areas to mark plot forsale
or notforsale are no longer limited to a town/residents availble claims.

Fixes resident claim limit not accounting for townblocks already
claimed.

Fixes /ta unclaim passing in a null TownBlockOwner meaning the available
value was left at 0.
  - Closes #5606.