Support for y less than 0 for revert-on-unclaim.
Add support for plotsnapshots made with MC 1.19 and newer, saving
plotblockdata below y level 1.

Snapshots now save down to the world bottom.

Older snapshots are loaded and have a min height of 0 set, which allows
them to be regenerated with the same code as newer snapshots.

We now split up snapshots knowing the correct vertical range, which
allows us to find the correct start and end point for each x,z "slice"
of the snapshot.

PlotBlockData version has been "documented" and versions 3 and lower
will no longer be loaded, as these used id:data or a
type:data-to-material hack to work. This has resulted in many of the
switch statements being removed entirely.

PlotBlockData version 5 now stores a minHeight int, similar to the
height int. If we are loading a version 5 plot snapshot we pull the
height from it, otherwise we set a minHeight of 0 (as this is as low as
old snapshots recorded.)