Make sure we aren't paying negative amounts to a Nation bank account during tax time.
Bankrupt towns' balances would return a negative amount when a nation
had % tax enabled.

This called the payTo() function with a negative amount for the
collector which then hits Vault with a negative deposit (which we are
warned not to do.)

Solution:
Don't pay a nation tax if the amount is less than or equal to zero.

Also added is an override to BankAcount for canPayFromHoldings() that
tests if a town is able to pay with debt.

Closes #6540.