Fix isUpperCased failing on underscores
The isUpperCased test meant that we were constantly re-converting the
entity-expl-revert list (and maybe more).

This resulted in the CREEPER being changed into minecraft:creeper, which
is not a valid option in our Entity registry. This resulted in the
CREEPER being removed from the list of entities whose explosions we
would revert.

What is still not working, and I have not had enough time to diagnose:

The config migration results I added has resulted in the world's
PlotManagementWildRegenEntities being wiped, and I can see no evidence
of the worldAction being done at all.