Fix overriding use-item and interact-with-block PlayerInteractEvent.
There are occaisions where plugins will cancel one of the above aspects
of the PlayerInteractEvent without cancelling the event as a whole.

Because we were specifically calling for the event to be uncancelled we
were overriding previously set use-item and interact-with-block
decisions made by other plugins.

This PR has Towny making setUseItemInHand and setUseInteractedBlock
decisions rather than fully uncancelling/cancelling events.

Closes #6861.