Fix removing Town from residents not being able to delete 1 man towns.
The issue presented itself when the Resident purge was set to
remove-town-only, which resulted in Towns not being removed when their
last resident is removed.

Resident#removeTown() was set to ignore the EmptyTownException, which
we're now listening to in order to remove the Town when the town becomes
empty.

Additionally:

- the DailyTimerTask has had a couple of tests added to make sure we're
not parsing over since-removed Towns when factoring conquered-days or
deleting 0 plot towns.

- removing a Resident from the datasource can no longer save a
non-existent Town which has that Resident as an outlaw or trusted
resident.

- removing a Nation can no longer cause a Town that no longer exists in
the Universe to save itself.

- the OnPlayerLogin task is no longer saving the default Town when it
adds a new Resident.