Enhance Nation Requires Proximity Feature
Optionally, towns in a nation can grant nearby towns the ability to
bypass the proximity-to-the-nation-capital value.

Also Optionally, there can be a hard-cap on how far a town can be from
the nation capital, even when a nation-town is nearby, preventing
nations going "wide."

Renames/moves the config settings into the Global Nation Settings
section & migrates old configs' settings.

Adds a new ProximityUtil where all the heavy lifting is done.

Spruced up the code in the various methods that used the proximity
settings.