Re-jiggers things to use recursion in order remove towns correctly.
Towns parsed over early could have been allowed to stay in the nation
because of a town which isn't removed until later.

This uses a while loop that runs until no further towns are removed, to
assure that the nation doesn't end up cut into two or more pieces
because of interior-towns being removed, while incorrectly allowing
far-away towns to remain in the nation.