Make Towny more resiliant to errors which come from having no Economy present on the server.
Alter ConfirmationTransaction to no longer take in an Account in the
constructor, instead it takes in a TownyObject, which will resolve an
account. This prevents errors when non-Economy-enabled servers attempt
to create a Confirmation with a cost set on it.

Make TownyEconomyHandler#getFormattedBalance able to handle returning a
formatted balance when the Economy is not active.

Add prettyMoney method to BaseCommand, use it throughout the command
classes to make things look cleaner.