Add /town cede plot {townname} command.
Players are tested to make sure that the town is not giving away their
homeblock, and that the other town is able to claim the townblock based
on all of Towny's rules for claiming, proximity, limits, ect.

Both giving and accepting players must accept Confirmations before the
land will change hands.

The command causes a TownPreClaimEvent to be thrown, which allows other
plugins to cancel the event.

Closes #7149.

  - New Command: /town cede plot {townname}
    - Attempts to give the townblock the player is stood in, to the
given town.

  - New Permission Node: towny.command.town.cede.plot
    - Child Node of towny.command.town.cede.* which is a child node of
towny.command.town.*
    - No changes required to the townyperms.yml