Fix nation deleted message showing when a predeletenationevent is cancelled.
Not really happy with the implementation of this, as it changes our
DataSource pattern to one where only a single removeObject method
returns a boolean. The TownUtil change is also not ideal now or before
this PR as the Nation can be prevented from being deleted, while still
processing a refund.

Potential alternative solutions:
- Run the message in the DatabaseHandler method, and lose the custom
deleted by admin message sent to the admin who deleted the nation.
- Run the PreDeleteNation event before the removeNation(Nation) in order
to prevent the TownUtil method being made redundant, and causing issues.

Closes #7380.