Update Resident Modes system.
Adds the ability to assign a permission node to a mode, and when
present, will require the player to have the permission node to toggle a
node.

This permission node is not required if the mode is part of a player's
default modes set via the permission plugin.

Additionally, the ability to clear a residents modes via command now
requires a permission node.

All together this makes it possible for admins to set modes that players
cannot remove.

Adds a ResidentModesInitializeEvent with which plugins can add Resident
Modes.

Uses a AbstractResidentMode class to allow for different types of Modes
to be made, which makes it possible for Modes to be exclusive of each
other: ie: you can no longer have townclaim and townunclaim modes active
at the same time.

New Permissions:
- towny.command.resident.set.mode.clear
- towny.command.resident.toggle.bedspawn
- towny.command.resident.toggle.bordertitles
- towny.command.resident.toggle.ignoreotherchannels
- towny.command.resident.toggle.infotool
- towny.command.resident.toggle.plotgroup
- towny.command.resident.toggle.townborder
- towny.command.resident.toggle.townunclaim

Additionally, the tab completer for toggling and setting modes should no
longer fall out of date, as ResidentModes will automatically get added
to the ResidentCommand tabcompleters.